* Removed duplicate function "function exists" * Eliminate throwing exception from DB function call * Add more validations before calling clone_schema * This is now baseline for passing all unit tests before putting in specific unit tests for clone_schema
