+++ koku/koku/test_migration_sql_helpers.py
@@ -44,25 +44,3 @@ class TestMigrationSQLHelpers(IamTestCase):
                 msh.apply_sql_file(None, filename)
         finally:
             os.unlink(filename)
-
-    def test_function_not_exists(self):
-        """
-        Test datbase function not found
-        """
-        res = msh.function_exists("public", "___no_func_here___", "public.___no_func_here___(eek text, ook text)")
-        self.assertFalse(res)
-
-    def test_function_exists(self):
-        with conn.cursor() as cur:
-            cur.execute(
-                """
-create function public.__eek(param1 text) returns text as $BODY$
-begin
-    return param1;
-end;
-$BODY$ language plpgsql;"""
-            )
-            res = msh.function_exists("public", "__eek", "public.__eek(param1 text)")
-            cur.execute("""drop function public.__eek(text);""")
-
-        self.assertTrue(res)
